+++ src/org/rapla/storage/xml/RaplaMainReader.java
@@ -92,7 +92,7 @@ public class RaplaMainReader extends RaplaXMLReader
             String repositoryVersion = atts.getValue("repositoryVersion");
             if ( repositoryVersion != null)
             {
-                resolver.setRepositoryVersion( Long.parseLong( repositoryVersion));
+                resolver.setRepositoryVersion( Integer.parseInt( repositoryVersion));
             }
             String start = atts.getValue( "startDate");
             String end = atts.getValue( "endDate");
